Prefetch from certificate stream (fixes #127)
Works around a bug in Conscrypt. The certificate stream is not read
there if InputStream.available() returns 0, which is the case in acme4j
since the stream is directly read from the CA via HTTP.

The workaround uses a BufferedInputStream and prefetches a few bytes
from the HTTP stream if available() is invoked.